backend/requirements.txt
click==6.6
Flask==0.11.1
Flask-Cors==3.0.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.3.1
six==1.10.0
Werkzeug==0.11.10
